# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Enom API",
        "description": "Enom REST API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.enom.ai",
            "description": "Production server"
        },
        {
            "url": "http://localhost/enom/public",
            "description": "Local XAMPP server"
        }
    ],
    "paths": {
...
        "/api/user/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get authenticated user profile",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "male",
                                                    "nullable": true
                                                },
                                                "dob": {
                                                    "type": "string",
                                                    "example": "1995-06-15",
                                                    "nullable": true
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Hello!",
                                                    "nullable": true
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "New York",
                                                    "nullable": true
                                                },
                                                "profile_image": {
                                                    "type": "string",
                                                    "example": "profile-images/photo.jpg",
                                                    "nullable": true
                                                },
                                                "is_verified": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update authenticated user profile",
                "description": "All fields are optional. Send only the fields you want to update. Use multipart/form-data when uploading a profile image.",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ]
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-06-15"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Hello, I am John."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token as: Bearer {token}",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}